+++ LethalThings/Thunderstore/CHANGELOG.md
@@ -1,12 +1,30 @@
 # Lethal Things 0.9.0
+**Content**
+- Added Shop Item: Pinger
+- Added Scrap: Gnarpy Plush
+
 **Systems**
 - Completely rewrite of Content Loader and Config system.
-- Config files are now synced between host and client.
+- Config files are now synced between host and client.  
+- Changed how utility belt adds slow for potential compatibility improvement.  
 
 **Bugfixes**
-- Boomba improvements.
-- Power outlet stun damage config is now working.
-- Added null check to tracking point which should fix one of the dartboard errors.
+- Lethal Things content should now spawn on modded moons  
+- Boomba improvements.  
+- Power outlet stun damage config is now working.  
+- Added null check to tracking point which should fix one of the dartboard errors.  
+- Fixed flare desync  
+- Rocket launcher laser turns off when pocketed.  
+- Evil maxwell desync.  
+- Flare gun client aim issue.  
+- Remote radar was able to switch target while turned off.  
+- Toy hammer was broken because of lethal company v47 update.  
+
+**Changes**
+- Toyhammer is now a scrap item by default  
+	- Added config option to add Toyhammer to store.  
+	- Added config option to make Toyhammer spawn as scrap.  
+- Cookie is now throwable.  
 
 # Lethal Things 0.8.8
 - Glizzy is no longer called "Training Manual" (idk how the heck that happened because it was called glizzy before.)